+++ packages/client/src/modals/backendError/index.tsx
@@ -1,28 +1,18 @@
-import { LogItem } from '@metastable/types';
-import React, { useState } from 'react';
+import React from 'react';
 
-import { API, TRPC } from '$api';
+import { API } from '$api';
 import { Button } from '$components/button';
 import { Log } from '$components/log';
 import { Modal, ModalActions, useModal } from '$components/modal';
+import { mainStore } from '$stores/MainStore';
 
 export const BackendError: React.FC = () => {
   const { close } = useModal();
-  const [log, setLog] = useState<LogItem[]>([]);
-
-  TRPC.instance.onBackendLog.useSubscription(undefined, {
-    onData: items => {
-      // TODO: trpc bug where the wrong response is being sent?
-      if (Array.isArray(items)) {
-        setLog(current => [...current, ...items]);
-      }
-    },
-  });
 
   return (
     <Modal title="Backend error" size="small">
       <div>Unable to start backend. Details:</div>
-      <Log items={log} />
+      <Log items={mainStore.backendLog} />
       <ModalActions>
         <Button
           variant="secondary"